feat(KNO-10034): create .knockbranch file when it does not exist (#624)
### Description

This PR updates the `knock branch switch` command so that when a `.knock_branch` file is not found, we prompt the user to create one automatically. We’ll look for an existing `.gitignore` file and create the `.knock_branch` file sibling to it. The directory containing `.gitignore` is our best guess for the root of the user’s project.

I also updated `knock branch exit` so that instead of deleting the `.knock_branch` file, it writes an empty string to it. This feels less intrusive than constantly deleting and re-creating `.knock_branch`.

### Tasks

[KNO-10034](https://linear.app/knock/issue/KNO-10034/cli-create-knock-branch-file-when-it-does-not-exist)

## Loom demo

https://www.loom.com/share/aaa825700da144ab8f6ea08208870860?sid=455ace12-9bf9-4d68-8137-93a0c4f388fa